refactor calendar event fetching to improve performance and structure
